requirements.txt
#Flask==2.2.2
#flask_restx==0.5.1
python-decouple==3.6

#werkzeug==2.2.2
#Werkzeug <= 2.1.2   
#flask == 2.1.2   
#flask-restx >= 0.5.1  
#werkzeug==2.1.2

#Flask-Login 0.6.2 

Werkzeug <= 2.1.2   
flask == 2.1.2   
flask-restx >= 0.5.1  